+++ CONASTEM/DataBase/views/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render
 
 # Create your views here.
 
-from ..models.Infraestructura.Equidad import Equidad
+from ..models.INFRAESTRUCTURA.Equidad import Equidad
 from ..models.Criterio import Criterio
 
 def index (request):
